pyproject.toml
[tool.poetry]
name = "blogpy"
version = "0.1.0"
description = ""
authors = ["Victor Hugo <victorblog410@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.12.4"
Django = "^3.2.4"
python-decouple = "^3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
